% Generated by roxygen2 (4.0.2): do not edit by hand
\name{advancedArg}
\alias{advancedArg}
\alias{advanced_arg}
\title{List advanced arguments}
\usage{
advancedArg(fun, package = "derfinder", browse = interactive())
}
\arguments{
\item{fun}{The name of a function(s) that has advanced arguments in
\code{package}.}

\item{package}{The name of the package where the function is stored. Only
'derfinder' and 'derfinderPlot' are accepted.}

\item{browse}{Whether to open the URLs in a browser.}
}
\value{
A list of URLs with the GitHub search queries.
}
\description{
Find in GitHub the documentation for the advanced arguments of a given
function.
}
\author{
Leonardo Collado-Torres
}

