% Generated by roxygen2 (4.0.2): do not edit by hand
\name{calculateStats}
\alias{calculateStats}
\alias{calculate_stats}
\title{Calculate F-statistics at base pair resolution from a loaded BAM files}
\usage{
calculateStats(coveragePrep, models, mc.cores = getOption("mc.cores", 1L),
  mc.outfile = Sys.getenv("SGE_STDERR_PATH"), adjustF = 0,
  lowMemDir = NULL, method = "Matrix", scalefac = 32, verbose = TRUE)
}
\arguments{
\item{coveragePrep}{A list with \code{$coverageProcessed},
\code{$mclapplyIndex}, and \code{$position} normally generated using
\link{preprocessCoverage}.}

\item{models}{A list with \code{$mod} and \code{$mod0} normally generated
using \link{makeModels}.}

\item{mc.cores}{This argument is passed to \link[BiocParallel]{SnowParam}
to define the number of \code{workers} used for running
\link[derfinderHelper]{fstats.apply}.}

\item{mc.outfile}{This argument is passed to \link[BiocParallel]{SnowParam}
to specify the \code{outfile} for any output from the workers.}

\item{adjustF}{A single value to adjust that is added in the denominator of
the F-stat calculation. Useful when the Residual Sum of Squares of the
alternative model is very small.}

\item{lowMemDir}{The directory where the processed chunks are saved when
using \link{preprocessCoverage} with a specified \code{lowMemDir}.}

\item{method}{This argument is passed to
\link[derfinderHelper]{fstats.apply}. Check the details there for more
information.}

\item{scalefac}{This argument is passed to
\link[derfinderHelper]{fstats.apply} and should be the same as the one used
in \link{preprocessCoverage}.}

\item{verbose}{If \code{TRUE} basic status updates will be printed along the
way.}
}
\value{
A numeric Rle with the F-statistics per base pair that passed the
cutoff.
}
\description{
After defining the models of interest (see \link{makeModels}) and
pre-processing the data (see \link{preprocessCoverage}), use
\link{calculateStats} to calculate the F-statistics at base-pair resolution.
}
\examples{
## Collapse the coverage information
collapsedFull <- collapseFullCoverage(list(genomeData$coverage),
    verbose=TRUE)

## Calculate library size adjustments
sampleDepths <- sampleDepth(collapsedFull, probs=c(0.5), nonzero=TRUE,
    verbose=TRUE)

## Build the models
group <- genomeInfo$pop
adjustvars <- data.frame(genomeInfo$gender)
models <- makeModels(sampleDepths, testvars=group, adjustvars=adjustvars)

## Preprocess the data
prep <- preprocessCoverage(genomeData, cutoff=0, scalefac=32, chunksize=1e3,
    colsubset=NULL)

## Run the function
fstats <- calculateStats(prep, models, mc.cores=1, verbose=TRUE,
    method='regular')
fstats

\dontrun{
## Compare vs pre-packaged F-statistics
summary(fstats - genomeFstats)
}
}
\author{
Leonardo Collado-Torres
}
\seealso{
\link{makeModels}, \link{preprocessCoverage}
}

