% Generated by roxygen2 (4.0.1): do not edit by hand
\name{createBw}
\alias{createBw}
\alias{create_bw}
\title{Export coverage to BigWig files}
\usage{
createBw(fullCov, path = ".", keepGR = TRUE,
  mc.cores = getOption("mc.cores", 1L),
  mc.outfile = Sys.getenv("SGE_STDERR_PATH"), verbose = TRUE)
}
\arguments{
\item{fullCov}{A list where each element is the result from
\link{loadCoverage} used with \code{returnCoverage = TRUE}. Can be generated
using \link{fullCoverage}.}

\item{path}{The path where the BigWig files will be created.}

\item{keepGR}{If \code{TRUE}, the \link[GenomicRanges]{GRanges} objects
created by \link{coerceGR} grouped into a \link[GenomicRanges]{GRangesList}
are returned. Otherwise they are discarded.}

\item{mc.cores}{This argument is passed to \link[BiocParallel]{SnowParam}
to define the number of \code{workers}. Use at most one core per chromosome.}

\item{mc.outfile}{This argument is passed to \link[BiocParallel]{SnowParam}
to specify the \code{outfile} for any output from the workers.}

\item{verbose}{If \code{TRUE} basic status updates will be printed along the
way.}
}
\value{
If \code{keepGR = TRUE}, then a \link[GenomicRanges]{GRangesList}
with the output for \link{coerceGR} for each of the samples.
}
\description{
Using output from \link{fullCoverage}, export the coverage from all the
samples to BigWig files using \link{createBwSample}.
}
\examples{
\dontrun{
}
}
\author{
Leonardo Collado-Torres
}
\seealso{
\link[GenomicRanges]{GRangesList}, \link[rtracklayer]{export},
\link{createBwSample}, \link{coerceGR}
}

