% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/Deprecated.R
\name{derfinder-deprecated}
\alias{advanced_arg}
\alias{analyze_chr}
\alias{annotate_regions}
\alias{calculate_pvalues}
\alias{calculate_stats}
\alias{coerce_gr}
\alias{collapse_full_coverage}
\alias{coverage_to_exon}
\alias{create_bw}
\alias{create_bw_sample}
\alias{derfinder-deprecated}
\alias{extended_map_seqlevels}
\alias{filter_data}
\alias{find_regions}
\alias{full_coverage}
\alias{get_region_coverage}
\alias{load_coverage}
\alias{make_genomic_state}
\alias{make_models}
\alias{merge_results}
\alias{preprocess_coverage}
\alias{raw_files}
\alias{region_matrix}
\alias{sample_depth}
\title{Deprecated functions in package \sQuote{derfinder}}
\usage{
advanced_arg()

analyze_chr()

annotate_regions()

calculate_pvalues()

calculate_stats()

coerce_gr()

collapse_full_coverage()

coverage_to_exon()

create_bw()

create_bw_sample()

extended_map_seqlevels()

filter_data()

find_regions()

full_coverage()

get_region_coverage()

load_coverage()

make_genomic_state()

make_models()

merge_results()

preprocess_coverage()

raw_files()

region_matrix()

sample_depth()
}
\description{
These functions are provided for compatibility with older versions
of \sQuote{derfinder} only, and will be defunct at the next release.
}
\details{
The following functions are deprecated and will be made defunct; use
 the replacement indicated below:
 \itemize{

   \item{advanced_arg: \code{\link{advancedArg}}}
   \item{analyze_chr: \code{\link{analyzeChr}}}
   \item{annotate_regions: \code{\link{annotateRegions}}}
   \item{calculate_pvalues: \code{\link{calculatePvalues}}}
   \item{calculate_stats: \code{\link{calculateStats}}}
   \item{coerce_gr: \code{\link{coerceGR}}}
   \item{collapse_full_coverage: \code{\link{collapseFullCoverage}}}
   \item{coverage_to_exon: \code{\link{coverageToExon}}}
   \item{create_bw: \code{\link{createBw}}}
   \item{create_bw_sample: \code{\link{createBwSample}}}
   \item{extended_map_seqlevels: \code{\link{extendedMapSeqlevels}}}
   \item{filter_data: \code{\link{filterData}}}
   \item{find_regions: \code{\link{findRegions}}}
   \item{full_coverage: \code{\link{fullCoverage}}}
   \item{get_region_coverage: \code{\link{getRegionCoverage}}}
   \item{load_coverage: \code{\link{loadCoverage}}}
   \item{make_genomic_state: \code{\link{makeGenomicState}}}
   \item{make_models: \code{\link{makeModels}}}
   \item{merge_results: \code{\link{mergeResults}}}
   \item{preprocess_coverage: \code{\link{preprocessCoverage}}}
   \item{raw_files: \code{\link{rawFiles}}}
   \item{region_matrix: \code{\link{regionMatrix}}}
   \item{sample_depth: \code{\link{sampleDepth}}}

 }
}

