\docType{package}
\name{derfinder2-package}
\alias{derfinder2-package}
\title{Fast differential expression analysis of RNA-seq data at base-pair resolution.}
\description{
  Fast differential expression analysis of per-nucleotide
  coverage tables.
}
\details{
  To do
}
\author{
  Leonardo Collado-Torres <lcollado@jhsph.edu>
}
\references{
  Paper coming soon.
}
\keyword{package}

