% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/Deprecated.R
\name{full_coverage}
\alias{full_coverage}
\title{Deprecated functions in package \sQuote{derfinder}}
\usage{
full_coverage()
}
\description{
These functions are provided for compatibility with older versions
of \sQuote{derfinder} only, and will be defunct at the next release.
}
\details{
The following functions are deprecated and will be made defunct; use
 the replacement indicated below:
 \itemize{

   \item{advanced_arg: \code{\link{advancedArg}}}
   \item{analyze_chr: \code{\link{analyzeChr}}}
   \item{annotate_regions: \code{\link{annotateRegions}}}
   \item{calculate_pvalues: \code{\link{calculatePvalues}}}
   \item{calculate_stats: \code{\link{calculateStats}}}
   \item{coerce_gr: \code{\link{coerceGR}}}
   \item{collapse_full_coverage: \code{\link{collapseFullCoverage}}}
   \item{coverage_to_exon: \code{\link{coverageToExon}}}
   \item{create_bw: \code{\link{createBw}}}
   \item{create_bw_sample: \code{\link{createBwSample}}}
   \item{extended_map_seqlevels: \code{\link{extendedMapSeqlevels}}}
   \item{filter_data: \code{\link{filterData}}}
   \item{find_regions: \code{\link{findRegions}}}
   \item{full_coverage: \code{\link{fullCoverage}}}
   \item{get_region_coverage: \code{\link{getRegionCoverage}}}
   \item{load_coverage: \code{\link{loadCoverage}}}
   \item{make_genomic_state: \code{\link{makeGenomicState}}}
   \item{make_models: \code{\link{makeModels}}}
   \item{merge_results: \code{\link{mergeResults}}}
   \item{preprocess_coverage: \code{\link{preprocessCoverage}}}
   \item{raw_files: \code{\link{rawFiles}}}
   \item{region_matrix: \code{\link{regionMatrix}}}
   \item{sample_depth: \code{\link{sampleDepth}}}

 }
}

