\docType{data}
\name{genomeFstats}
\alias{genomeFstats}
\title{F-statistics for the example data}
\format{A numeric Rle of length 1434 with the calculated F-statistics as
exemplified in \link{calculateStats}.}
\description{
Calculated F-statistics for a 10kb region from chr21
processed for 31 RNA-seq samples described in
\link{genomeInfo}. For more information check the example
of \link{calculateStats}.
}
\seealso{
\link{calculateStats}
}
\keyword{datasets}

