% Generated by roxygen2 (4.0.2): do not edit by hand
\docType{data}
\name{genomicState}
\alias{genomicState}
\title{Genomic State for Hsapiens.UCSC.hg19.knownGene}
\format{A GRangesList with two components.
\describe{
\item{fullGenome }{ classifies each region as either being exon, intron or
intragenic.}
\item{codingGenome }{ classfies the regions as being promoter, exon, intro,
5UTR, 3UTR or intragenic.}
}}
\description{
Pre-computed genomic state for Hsapiens UCSC hg19 knownGene annotation built
using \link[derfinder]{makeGenomicState} for
TxDb.Hsapiens.UCSC.hg19.knownGene version 2.14.0. The object has been subset
for chr21 only.
}
\seealso{
\link[derfinder]{makeGenomicState}
}
\keyword{datasets}

