\name{loccov2}
\alias{loccov2}
%- Also NEED an '\alias' for EACH other topic documented here.
\title{ helper function for locfdrFit
}
\description{
used only internally in \code{locfdrFit}, equivalent to \code{locfdr:::loccov2}
}
\usage{
loccov2(X, X0, i0, f, ests, N)
}
\details{
This is an internal function of the \code{locfdr} package, so documentation is not available. 
}
\author{
Bradley Efron
}


\seealso{
\code{\link{locfdr}}
}
