\name{locmle}
\alias{locmle}
%- Also NEED an '\alias' for EACH other topic documented here.
\title{ helper function for locfdrFit
}
\description{
used only internally in \code{locfdrFit}, equivalent to \code{locfdr:::locmle}
}
\usage{
locmle(z, xlim, Jmle = 35, d = 0, s = 1, ep = 1/1e+05, sw = 0, Cov.in)
}
\details{
This is an internal function of the \code{locfdr} package, so documentation is not available. 
}
\author{
Bradley Efron
}
\seealso{
\code{\link{locfdr}}
}