% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/advancedArg.R
\name{advancedArg}
\alias{advancedArg}
\title{List advanced arguments}
\usage{
advancedArg(fun, package = "derfinder", browse = interactive())
}
\arguments{
\item{fun}{The name of a function(s) that has advanced arguments in
\code{package}.}

\item{package}{The name of the package where the function is stored. Only
'derfinder', 'derfinderPlot', and 'regionReport' are accepted.}

\item{browse}{Whether to open the URLs in a browser.}
}
\value{
A vector of URLs with the GitHub search queries.
}
\description{
Find in GitHub the documentation for the advanced arguments of a given
function.
}
\details{
If you are interested on the default options used for functions
that run on multiple cores, check
https://github.com/lcolladotor/derfinder/blob/master/R/utils.R
Note that in general, \link[BiocParallel]{SnowParam} is more memory efficient
than link[BiocParallel]{MulticoreParam}. If you so desire, use your favorite
cluster type by specifying \code{BPPARAM}.
}
\examples{
## Open the advanced argument docs for loadCoverage()
if(interactive()) {
    advancedArg('loadCoverage')
} else {
    (advancedArg('loadCoverage', browse = FALSE))
}
}
\author{
Leonardo Collado-Torres
}

