% Generated by roxygen2 (4.0.2): do not edit by hand
\name{calculateStats}
\alias{calculateStats}
\alias{calculate_stats}
\title{Calculate F-statistics at base pair resolution from a loaded BAM files}
\usage{
calculateStats(coveragePrep, models, lowMemDir = NULL, ...)
}
\arguments{
\item{coveragePrep}{A list with \code{$coverageProcessed},
\code{$mclapplyIndex}, and \code{$position} normally generated using
\link{preprocessCoverage}.}

\item{models}{A list with \code{$mod} and \code{$mod0} normally generated
using \link{makeModels}.}

\item{lowMemDir}{The directory where the processed chunks are saved when
using \link{preprocessCoverage} with a specified \code{lowMemDir}.}

\item{...}{Arguments passed to other methods and/or advanced arguments.}
}
\value{
A numeric Rle with the F-statistics per base pair that passed the
cutoff.
}
\description{
After defining the models of interest (see \link{makeModels}) and
pre-processing the data (see \link{preprocessCoverage}), use
\link{calculateStats} to calculate the F-statistics at base-pair resolution.
}
\examples{
## Collapse the coverage information
collapsedFull <- collapseFullCoverage(list(genomeData$coverage),
    verbose = TRUE)

## Calculate library size adjustments
sampleDepths <- sampleDepth(collapsedFull, probs=c(0.5), verbose = TRUE)

## Build the models
group <- genomeInfo$pop
adjustvars <- data.frame(genomeInfo$gender)
models <- makeModels(sampleDepths, testvars = group, adjustvars = adjustvars)

## Preprocess the data
prep <- preprocessCoverage(genomeData, cutoff = 0, scalefac = 32,
    chunksize=1e3, colsubset=NULL)

## Run the function
fstats <- calculateStats(prep, models, verbose=TRUE, method='regular')
fstats

\dontrun{
## Compare vs pre-packaged F-statistics
library('testthat')
expect_that(fstats, is_equivalent_to(genomeFstats))
}
}
\author{
Leonardo Collado-Torres
}
\seealso{
\link{makeModels}, \link{preprocessCoverage}
}

