% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Deprecated.R
\name{derfinder-deprecated}
\alias{advancedArg}
\alias{derfinder-deprecated}
\title{Deprecated functions in package \sQuote{derfinder}}
\usage{
advancedArg(fun, package = "derfinder", browse = interactive())
}
\arguments{
\item{fun}{The name of a function(s) that has advanced arguments in
\code{package}.}

\item{package}{The name of the package where the function is stored. Only
'derfinder', 'derfinderPlot', and 'regionReport' are accepted.}

\item{browse}{Whether to open the URLs in a browser.}
}
\value{
A vector of URLs with the GitHub search queries.
}
\description{
These functions are provided for compatibility with older version of
\sQuote{derfinder} only and will be defunct at the next release.
}
\details{
The following functions are deprecated and will be made defunct.

\itemize{
    \item{advancedArg}{ Not needed now that the advanced arguments are
    documented on the help pages of each function.}
}
}
\examples{
## Open the advanced argument docs for loadCoverage()
if(interactive()) {
    advancedArg('loadCoverage')
} else {
    (advancedArg('loadCoverage', browse = FALSE))
}

}

