% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/derfinder-package.R
\docType{package}
\name{derfinder-package}
\alias{derfinder}
\alias{derfinder-package}
\title{derfinder: Annotation-agnostic differential expression analysis of RNA-seq data at
    base-pair resolution via the DER Finder approach}
\description{
This package provides functions for annotation-agnostic
    differential expression analysis of RNA-seq data. Two implementations of
    the DER Finder approach are included in this package: (1) single base-level
    F-statistics and (2) DER identification at the expressed regions-level.
    The DER Finder approach can also be used to identify differentially bounded
    ChIP-seq peaks.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/lcolladotor/derfinder}
  \item Report bugs at \url{https://support.bioconductor.org/t/derfinder/}
}

}
\author{
\strong{Maintainer}: Leonardo Collado-Torres \email{lcolladotor@gmail.com} (\href{https://orcid.org/0000-0003-2140-308X}{ORCID})

Authors:
\itemize{
  \item Andrew E. Jaffe \email{andrew.jaffe@libd.org} (\href{https://orcid.org/0000-0001-6886-1454}{ORCID})
  \item Jeffrey T. Leek \email{jtleek@gmail.com} (\href{https://orcid.org/0000-0002-2873-2671}{ORCID}) [thesis advisor]
}

Other contributors:
\itemize{
  \item Alyssa C. Frazee \email{alyssa.frazee@gmail.com} [contributor]
}

}
\keyword{internal}
