% Generated by roxygen2 (4.0.1): do not edit by hand
\name{getRegionCoverage}
\alias{getRegionCoverage}
\title{Extract coverage information for a set of regions}
\usage{
getRegionCoverage(fullCov, regions, totalMapped = NULL, targetSize = 8e+07,
  mc.cores = 1, mc.outfile = Sys.getenv("SGE_STDERR_PATH"),
  chrsStyle = "UCSC", verbose = TRUE)
}
\arguments{
\item{fullCov}{A list where each element is the result from
\link{loadCoverage} used with \code{returnCoverage = TRUE}. Can be generated
using \link{fullCoverage}.}

\item{regions}{The \code{$regions} output from \link{calculatePvalues}. It
is important that the seqlengths information is provided.}

\item{totalMapped}{The total number of reads mapped for each sample.
Providing this data adjusts the coverage to reads in \code{targetSize}
library. By default, to reads per 80 million reads.}

\item{targetSize}{The target library size to adjust the coverage to. Used
only when \code{totalMapped} is specified.}

\item{mc.cores}{This argument is passed to \link[BiocParallel]{SnowParam}
to define the number of \code{workers}. You should use at most one core per
chromosome.}

\item{mc.outfile}{This argument is passed to \link[BiocParallel]{SnowParam}
to specify the \code{outfile} for any output from the workers.}

\item{chrsStyle}{The naming style of the chromosomes. By default, UCSC. See
\link[GenomeInfoDb]{seqlevelsStyle}.}

\item{verbose}{If \code{TRUE} basic status updates will be printed along the
way.}
}
\value{
a list of data.frame where each data.frame has the coverage
information (nrow = width of region, ncol = number of samples) for a given
region. The names of the list correspond to the region indexes in
\code{regions}
}
\description{
This function extracts the raw coverage information calculated by
\link{fullCoverage} at each base for a set of regions found with
\link{calculatePvalues}. It can further calculate the mean coverage per
sample for each region.
}
\details{
When \code{fullCov} is the output of \link{loadCoverage} with
\code{cutoff} non-NULL, \link{getRegionCoverage} assumes that the regions
come from the same data. Meaning that \link{filterData} was not used again.
This ensures that the regions are a subset of the data available in
\code{fullCov}.
}
\examples{
## Obtain fullCov object
fullCov <- list('21'=genomeDataRaw$coverage)

## Assign chr lengths using hg19 information, use only first two regions
library('GenomicRanges')
data(hg19Ideogram, package = 'biovizBase', envir = environment())
regions <- genomeRegions$regions[1:2]
seqlengths(regions) <- seqlengths(hg19Ideogram)[names(seqlengths(regions))]

## Finally, get the region coverage
regionCov <- getRegionCoverage(fullCov=fullCov, regions=regions)
}
\author{
Andrew Jaffe, Leonardo Collado-Torres
}
\seealso{
\link{fullCoverage}, \link{calculatePvalues}
}

