% Generated by roxygen2 (4.0.2): do not edit by hand
\name{makeGenomicState}
\alias{makeGenomicState}
\alias{make_genomic_state}
\title{Obtain the genomic state per region from annotation}
\usage{
makeGenomicState(txdb, chrs = c(1:22, "X", "Y"), ...)
}
\arguments{
\item{txdb}{A \link[GenomicFeatures]{TxDb} object.}

\item{chrs}{The names of the chromosomes to use as denoted in the
\code{txdb} object. Check \link[GenomicFeatures]{isActiveSeq}.}

\item{...}{Arguments passed to other methods and/or advanced arguments.}
}
\value{
A \code{GRangesList} object with two elements: \code{fullGenome} and
\code{codingGenome}. Both have metadata information for the type of region
(theRegion), transcript IDs (tx_id), transcript name (tx_name), and gene ID
(gene_id). \code{fullGenome} classifies each region as either being exon,
intron or intragenic. \code{codingGenome} classfies the regions as being
promoter, exon, intro, 5UTR, 3UTR or intragenic.
}
\description{
This function summarizes the annotation contained in a
\link[GenomicFeatures]{TxDb} at each given base of the genome based
on annotated transcripts. It groups contiguous base pairs classified as the
same type into regions.
}
\examples{
## Load the example data base from the GenomicFeatures vignette
library('GenomicFeatures')
samplefile <- system.file('extdata', 'UCSC_knownGene_sample.sqlite',
    package='GenomicFeatures')
txdb <- loadDb(samplefile, dbType = 'TxDb', dbPackage = 'GenomicFeatures')

## Generate genomic state object, only for chr21
sampleGenomicState <- makeGenomicState(txdb, chrs='chr21')
\dontrun{
## Create the GenomicState object for Hsapiens.UCSC.hg19.knownGene
library('TxDb.Hsapiens.UCSC.hg19.knownGene')
txdb <- TxDb.Hsapiens.UCSC.hg19.knownGene

## Creating this GenomicState object takes around 8 min for all chrs and
## around 30 secs for chr21
GenomicState.Hsapiens.UCSC.hg19.knownGene.chr21 <-
    makeGenomicState(txdb=txdb, chrs='chr21')

## For convinience, this object is already included in derfinder
identical(GenomicState.Hsapiens.UCSC.hg19.knownGene.chr21, genomicState)

## Hsapiens ENSEMBL GRCh37
library('GenomicFeatures')
## Can take several minutes and speed will depend on your internet speed
xx <- makeTxDbPackageFromBiomart(version = '0.99', maintainer = 'Your Name',
    author='Your Name')
txdb <- loadDb(file.path('TxDb.Hsapiens.BioMart.ensembl.GRCh37.p11', 'inst',
    'extdata', 'TxDb.Hsapiens.BioMart.ensembl.GRCh37.p11.sqlite'))

## Creating this GenomicState object takes around 13 min
GenomicState.Hsapiens.ensembl.GRCh37.p11 <- makeGenomicState(txdb=txdb,
    chrs=c(1:22, 'X', 'Y'))

## Save for later use
save(GenomicState.Hsapiens.ensembl.GRCh37.p11,
    file='GenomicState.Hsapiens.ensembl.GRCh37.p11.Rdata')
}
}
\author{
Andrew Jaffe, Leonardo Collado-Torres
}
\seealso{
\link[GenomicFeatures]{TxDb}
}

