% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/regionMatrix.R
\name{regionMatrix}
\alias{regionMatrix}
\title{Identify regions data by a coverage filter and get a count matrix}
\usage{
regionMatrix(fullCov, cutoff = 5, filter = "mean", L, runFilter = TRUE,
  returnBP = TRUE, ...)
}
\arguments{
\item{fullCov}{A list where each element is the result from 
\link{loadCoverage} used with \code{returnCoverage = TRUE}. Can be generated 
using \link{fullCoverage}. If \code{runFilter = FALSE}, then 
\code{returnMean = TRUE} must have been used.}

\item{cutoff}{The base-pair level cutoff to use. It's behavior is controlled 
by \code{filter}.}

\item{filter}{Has to be either \code{'one'} (default) or \code{'mean'}. In 
the first case, at least one sample has to have coverage above \code{cutoff}.
In the second case, the mean coverage has to be greater than \code{cutoff}.}

\item{L}{The width of the reads used. Either a vector of length 1 or length
equal to the number of samples.}

\item{runFilter}{This controls whether to run \link{filterData} or not. If 
set to \code{FALSE} then \code{returnMean = TRUE} must have been used to 
create each element of \code{fullCov}.}

\item{returnBP}{If \code{TRUE}, returns \code{$bpCoverage} explained below.}

\item{...}{Arguments passed to other methods and/or advanced arguments.}
}
\value{
A list with one entry per chromosome. Then per chromosome, a list 
with three components.
\describe{
\item{regions }{ A set of regions based on the coverage filter cutoff as
returned by \link{findRegions}.}
\item{bpCoverage }{ A list with one element per region. Each element is a matrix with numbers of rows equal to the number of base pairs in the region and number of columns equal to the number of samples. It contains the base-level coverage information for the regions. Only returned when \code{returnBP = TRUE}.}
\item{coverageMatrix }{  A matrix with the mean coverage by sample for each
candidate region.}
}
}
\description{
Given a set of un-filtered coverage data (see \link{fullCoverage}), create
candidate regions by applying a cutoff on the coverage values,
and obtain a count matrix where the number of rows corresponds to the number
of candidate regions and the number of columns corresponds to the number of 
samples. The values are the mean coverage for a given sample for a given 
region.
}
\details{
This function uses several other \link{derfinder-package} 
functions. Inspect the code if interested.

You should use at most one core per chromosome.
}
\examples{
## Create some toy data
library('IRanges')
x <- Rle(round(runif(1e4, max=10)))
y <- Rle(round(runif(1e4, max=10)))
z <- Rle(round(runif(1e4, max=10)))
fullCov <- list('chr21' = DataFrame(x, y, z))

## Calculate a proxy of library size
libSize <- sapply(fullCov$chr21, sum)

## Run region matrix normalizing the coverage
regionMat <- regionMatrix(fullCov = fullCov, maxRegionGap = 10L, 
    maxClusterGap = 300L, L = 36, totalMapped = libSize, targetSize = 4e4)

\dontrun{
## You can alternatively use filterData() on fullCov to reduce the required
## memory before using regionMatrix(). This can be useful when mc.cores > 1
filteredCov <- lapply(fullCov, filterData, returnMean=TRUE, filter='mean', 
    cutoff=5, totalMapped = libSize, targetSize = 4e4)
regionMat2 <- regionMatrix(filteredCov, maxRegionGap = 10L, 
    maxClusterGap = 300L, L = 36, runFilter=FALSE)
identical(regionMat2, regionMat)
}
}
\author{
Leonardo Collado-Torres
}

